Separated questions data and helper (questions classes) into different files
